clevis-udisks2: enable dropping of privileges for clevis decrypt
We have to run as root in order to read the metadata from the disk.
However, we don't have to process that metadata as root. In fact,
it would be better not to. This patch enables the decryption itself
to be done as the specified user. This user CANNOT be the same user
as is running the process otherwise they could ptrace the processes
in order to recover the key.