Specify unprivileged user/group at compile time
The previous attempt to drop privileges unfortunately opened up a
security hole because the user could run the binary as any user
he wanted, including himself. The user could then run as himself
and then ptrace the clevis processes to obtain the key.

Instead, we specify the user and group at compile time.