Cleaned script a bit; Cleans up when finished.
 * Removed the vim settings (I don't use vim so I can't even check if I'm
sticking to its specifications, so why bother)

 * Moved things to sane
locations: makes it easier to see where things go in
/usr/share/initramfs-tools/

 * Included a script that does some
cleanup: If clevis is still running, it kills the process (For example
when the user enters the password before clevis can decrypt anything).
It also cleans up by flushing configuration from all interfaces.

 * Fixed locations for clevis and such in initramfs: also set up PATH
specifically to include clevis and curl. I just included 'what works':
this might be able to be trimmed down, but it doesn't hurt as it is now.