Add Trusted Platform Module 2.0 (TPM2) pin support
This adds support for a Trusted Platform Module 2.0 (TPM2) clevis pin.
It encrypts the JWK using a TPM2 chip, and then at decryption time the
JWK is decrypted using the same TPM2 chip to allow clevis to decrypt
the secret stored in the JWE. That way the ciphertext is bound to a TPM
so it can only be decripted in a particular machine.

Encrypting data using the tpm2 pin works the same than with other pins:

$ clevis encrypt tpm2 '{}' < PT > JWE

The pin has reasonable defaults for its configuration, but a different
hierarchy, hash, and key algorithms can be choosen if the defaults are
not suitable.

Decryption also works the same than with other pins, only needs the JWE:

$ clevis decrypt tpm2 < JWE > PT

The public and private key pair of the encrypted JWK are stored in the
JWE object, so those can be fetched at decryption time to unseal the
JWK using the TPM2 chip.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>