Allow specifying the user/group in udisks2
Our previous code disallowed running the executable in any scenario
except SETUID=root. However, this made debugging and automated testing
harder than it needs to be. We thus change the code so that if the
program is exected as root directly, the caller can specify the
user/group of the desktop user. This is not allowed in the SETUID case
to prevent running code as another user.