Don't attempt to create temporary files in the user's home directory
To use the tpm2 pin, some temporary files need to be created due how the
tpm2-tools work. Currently they are created in the user's home directory
but the commands can be executed by a user that doesn't have a home dir.

So it's better to just create the temporary directory in /tmp, which is
mktemp default. The mktemp default permissions are u+rwx anyways, so it
isn't less secure to have the temp dir at /tmp instead of the home dir.

Fixes: #30

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>