Don't execute clevis-luks-unlock while loop in a subshell
The loop that tries to open the dm-crypt devices using the pins in the
luksmeta header is executed in a subshell. So on success it calls exit
to exit the subshell.

But then clevis-luks-unlock has no way to know if the encrypted device
was opened correctly or not. So run the loop in the main shell process
and return 0 as exit status if the operation was successful.

Fixes: #36

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>