Try to fix race condition in systemd.path (https://github.com/systemd/systemd/issues/5770). When clevis-luks-askpass.path is triggered and starts clevis-luks-askpass.service it may happen that only one file is in /run/systemd/ask-password and only this file is being processed. All files created in the meantime are being ignored as clevis-luks-askpass.path won't queue to start clevis-luks-askpass.service again after it finished and if there are still files in /run/systemd/ask-password/.
Now clevis-luks-askpass checks if there are files left which were not previously processed before it exits.
There is still a small window to let this race condition happen, but this could not be avoided.

Also systemd's PasswordAgents do not have a return channel to determine if the provided password is correct.
Thus clevis-luks-askpass checks if a file in /run/systemd/ask-password/ was previously processed and ignores
it, because otherwhise it will end up in a loop.