clevis-luks-askpass: replace nc by ncat
nc is assumed to be ncat from Nmap for the --send-only option to work.
This assumption holds true on Fedora, where nc is a symbolic link to
ncat, while other distributions only ship the binary with the original
upstream name. Replacing the name makes it clearer which version of nc
is expected and improves compatibility with other distributions while
retaining compatibility with Fedora.