Speed up PBKDF
From cryptsetup docs:
" A PBKDF is used for increasing dictionary and brute-force attack cost for keyslot passwords."

Rationale: $key and LUKS Master Key have the same entropy. Therefore the cost of attack on each is about the same and we do not need to waste time in PBKDF. This speeds up mount times by ~2 seconds.