Validate the pcr_ids and error out if parsing fails
As of now, our parsing of pcr_ids fails if we try to use an actual JSON
array, and when it happens, we end up silently using no policy at all,
which is certainly not what the user requested/intended.

This patch improves our parsing to support reading both strings and JSON
arrays for the pcr_ids propery. If we encounter an error, we report it
and exit.

Fixes: #103