tpm2: fix pcr_ids validation
In PR#140 we improved the validation for pcr_ids, however we introduced
an error in this validation, which broke cases in which we specify
something in the configuration, but not the pcr_ids attribute.

E.g.: '{"key": "ecc"}' now fails with "Parsing the requested policy
failed!".

In this commit we fix this issue and add a new test case to cover it.