initramfs: Warn if it's not clear which nic to configure
If the user doesn't provide an ip= command line, configure_networking()
will try to configure each interface until one succeeds. But the one that
configures fastest may not be the one that can communicate with our server.
If we detect a situation where there are multiple configurable NICs and no
ip= parameter, emit a warning so that the user considers this as a possible
reason for failure.