luks: get rid of bashisms in clevis-luks-common-functions
As suggested in #202, let's try to define some sort of style guide
to try to improve portability and consistency across the scripts.

The suggestions include:

- getting rid of bashisms, at least for scripts running in early
  userland
- running with "set -e" and "set -u"
- using $() instead of backticks
- using long form of `jose' commands, for better readability

This commit converts clevis-luks-common-function and its user
clevis-luks-list.

As of now, shellcheck complains of the following:
"SC2039: In POSIX sh, 'local' is undefined."

This can be muted with `shellcheck -eSC2039'