systemd: drop hard requirement on networking
Whether we need networking or not for unlocking an encrypted block
device is a property of the block device in question. This is expressed
in `/etc/crypttab` via the `_netdev` option. For example, the systemd
cryptsetup generator[1] picks up on this and correctly orders unlocking
of devices that need networking after `remote-fs-pre.target`.

Thus, we shouldn't need to unconditionally require and order ourselves
after networking comes up. Let whatever interprets `/etc/crypttab` take
care of this.

Add `DefaultDependencies=no` because we need to be able to run well
before `sysinit.target`.

[1] https://www.freedesktop.org/software/systemd/man/systemd-cryptsetup-generator.html