luks: fix comparison of pin configutations in tests
pin_cfg_equal() was assigning the same value for both configurations
passed as parameters. Fix this and also remove the dependency on the
command diff, as we can compare the JSON value in the variables
directly.

This should also fix the CI failure in some platforms where we do not
have diff installed.