dracut: favour systemd units over dracut hooks
In general, when possible it's better to use systemd units instead of
dracut hooks because it's (not surprisingly) easier to integrate with
other systemd units in the initramfs.

Let's do our part towards eventually phasing out the "init system within
an init system" situation we're in right now and prefer systemd units on
systemd-based initrds.

It also mirrors how unlocking works in the real root.

You can see this similar "don't ship hooks on systemd" pattern in other
dracut modules (for example `04watchdog`, `50plymouth`, and `90crypt`).