Introduce -y (assume yes) argument to clevis luks bind
In order to simplify automated operations with e.g. ansible,
it would be helpful to have a way to automate the creation of
bindings with clevis.

In simple scenarios, it's possible to download the advertisement
from a tang server and pass it in the binding configuration, to
do the binding offline, in the following way:

curl -sfg http://tang.server/adv -o adv.jws

clevis luks bind -d /dev/sda2 tang '{"url":"http://tang.server", "adv":"adv.jws}'

However, for more complex scenarios using multiple servers with
the sss pin, it becomes a lot more complicated to do the same
thing and do the binding in an automated fashion. An alternative
would be to use expect (tcl), but it can also be complicated.

In this commit we introduce -y as a parameter meanining _assume yes_.
Essentially, this would make it so that the user would not have to
manually trust tang key(s) by typing y/yes.

Security-wise, it would be similar to downloading the advertisement
manually and passing it to tang as the "adv" configuration option,
something already supported, and the same caveat of making sure the
network isn't compromised applies here as well.