systemd: rework clevis-luks-askpass for improved reliability
clevis-luks-askpass has been refactored so that it becomes both
simpler and more reliable.

We now get the list of devices to be unlocked from crypttab, which
makes it simpler to verify whether there are any devices pending to
be unlocked. This improves the reliability in the situation when we
want to unlock multiple devices.

Also, remove the suggestion to add _netdev to crypttab/fstab, as that
is no longer required and in practice could be problematic in many
situations, as it would create dependencies from units to be mounted
during the boot process.

To set up multiple LUKS devices to be unlocked during the boot process,
do the following:
1) create clevis bindings for all the devices
2) run dracut -f to update the initramfs
3) enable clevis-luks-askpass.path unit (systemctl enable
   clevis-luks-askpass.path), so that devices that are not unlocked in
   early boot will be unlocked after switch-root.
   There is no harm in enabling this unit even if there are no devices
   to be unlocked after switch root, so it might be a good idea to
   simply enable it always.
4) if using tang, network needs to be setup, as since c52caeb4 (dracut:
   drop rd.neednet=1 injection), we do not add `rd.neednet=1`
   automatically anymore, in order to better support generic initrds
   and work similar to other root-on-{NFS,iSCSI,NBD,...} schemes, where
   one must explicitly configure networking, when required.
   If using DHCP, passing rd.neednet=1 -- via e.g. grub or dracut's
   --kernel-cmdline option -- should be enough.