systemd: clevis-luks-askpass: exit cleanly with SIGTERM
Especially when running in early boot, clevis-luks-askpass may be
looping because e.g. we still have devices remaining to unlock, as
per crypttab, or we were unable to determine if there are any devices
left to be unlocked.

Eventually we may receive a SIGTERM, from e.g. systemd and we should
exit cleanly in this situation.