initramfs: Make network configuration on-demand
Today we always try to configure the networking before we even know
if it is necessary. It's not necessary if, say, the volume is using a
TPM pin, or perhaps no pin at all and clevis just happens to be
installed. In those cases, the user is stuck waiting for
configure_networking() to finish before they get prompted for a passphrase,
and that can take nearly 5 minutes to timeout if the system is offline.

Let's move the configure_networking() code into clevisloop() and only
run it if we detect a LUKS volume using a tang pin. Since clevisloop()
is executing in the background, the boot can continue in the meantime
allowing the user to unlock via passphrase without first waiting for
network configuration to complete or timeout.

Since we use `clevis-luks-list` to detect the pin type, include it in
in the initramfs.