ci: use github actions
Add a simple workflow to perform the build and run the tests
using the following distros (the same we had with travis):

- fedora:rawhide
- fedora:latest
- centos:7
- centos:8
- debian:testing
- debian:latest
- ubuntu:devel
- ubuntu:rolling
- ubuntu:bionic

Github actions seems slightly faster than travis, not to mention that
its better container integration might show to be useful in terms of
reliability: often we have to re-run travis jobs that failed in the
step of container creation.

Once the CI looks good, we can remove travis afterwards.