ci: use travis for testing dracut unlocker
Since we already run the regular tests with Github actions, let's take
advantage of travis' nested virtualization capability and use it to test
the dracut unlocker, which is a more complex scenario, since it requires
a reboot. In practice, we will install a new VM and test boot unlocking.

This test is performed with the following:
- CentOS 8
- CentOS 8 stream
- Fedora 32
- Fedora rawhide