pins: tang - improve advertisement validation on bind
When performing a bind using the tang pin, we use an advertisement
either obtained from the tang server or passed by in the tang
configuration itself.

If this advertisement is not valid, i.e., not valid JSON, sometimes we
may leak out jose error messages in stderr, which makes the output very
confusing.

Example output would be the following:

Invalid json!
Usage: jose fmt [OPTIONS]

Converts JSON between serialization formats

-X --not Invert the following assertion

-O --object Assert TOP to be an object

-A --array Assert TOP to be an array

-S --string Assert TOP to be a string

-I --integer Assert TOP to be an integer

[...]

In this commit we improve the validation of the advertisement to be used
and also make sure we don't leak raw jose output in the error messages.