luks: replace `seq` with bash's builtin sequence expression
This gives the same result without the added dep.

Hit this in FCOS where we don't currently ship `seq` in the initramfs
(normally, the clevis dracut module should've installed it if it
requires it). We could do that of course, but it's even better if we can
avoid it entirely.