tang: default JWK thumbprint is now SHA-256 / deprecate SHA-1
During encryption using tang, the clevis metadata stores the thumbprint
of the exchange key in the `kid' attribute. This thumbprint used jose
default thumbprint hash algorithm -- SHA-1 as of now, even though
RFC 7638 recommends at least SHA-256.

During decryption with tang, we then validate the key with the thumbprint
represented by `kid' is present in the advertisement. At this point, we
used again jose default thumbprint hash for this check. We do have the
JWK, though, so we are able to recalculate the thumbprint with different
algorithms to make sure it matches.

clevis now moved to use SHA-256 for this thumbprint, and deprecated SHA-1
use. It still supports SHA-1 so that previously encrypted data can still
be decrypted, but a warning will be issued when decryption finds a
thumbprint using a deprecated hash.