test: define the key size for our tests (#320)
Older cryptsetup had a different default value for the key size,
e.g. 256, whereas newer ones have changed it to 512.

Let's specify the key size when creating our LUKS devices for the
tests, to make the devices more similar even when running in older
versions of cryptsetup.

We now use 512 as the key size for our test devices.