initramfs: add support for mkinitcpio initramfs generator
mkinitcpio is the default mkinitcpio used by Arch Linux. It can be extended
through custom hooks, so provide one that installs all the necessary files for
"clevis luks unlock" to function. The device to be unlocked is configured
through the "cryptdevice" kernel command line option in the same way the
built-in "encrypt" mkinitcpio hook works.