sss: sss-encrypt must read pipe data until EOF is reached
With current implementation sss-encrypt opens a pipe to the child pin,
then tries to read this pipe *only once*. This might return only partial
data or no data at all in case if the child pin response is slow.

Fix it by reading the pipe until it is closed. And only then process the
data.

Also increase the response buffer from 1024 bytes to 4096 bytes to
account for big responses that can be larger than 1024 bytes.

Fixes #389